+++ features/step_definitions/wallet_steps/wallet_homepage_steps.rb
@@ -4,39 +4,67 @@ Given(/^I verify I am on the wallet homepage$/) do
 end
 
 Given(/^I verify that the wallet home hero is displayed$/) do
-  expect(page).to have_css(".content-container",:text => "ID.me Wallet empowers you to manage all of the moving parts of your digital identity. With one login, you'll be able to easily access benefits and discounts from hundreds of partners. Start taking control of your digital identity today.")
-  expect(page).to have_css(".wallet-home-hero-cta a",:visible => true)
+  expect(page).to have_css(".page-container",:text => "ID.me simplifies how individuals securely prove and share their identity online.\nFor Business & Government keyboard_arrow_right\nFor Individuals keyboard_arrow_right")
+  expect(page).to have_css(".m_feature_image",:visible => true)
 end
 
-Given(/^I verify the wallet sign in links direct to idp$/) do
-  @WalletHomepage.click_shared_nav_sign_in_button
+Given("I verify the For Business and Government navigation dropdown link {string}") do |dropdown_link|
+  @WalletHomepage.click_shared_navlink_forbusiness
+  find("#js_dropdown1").click_link(dropdown_link)
+
+  expected_url, page_header = case dropdown_link
+                              when "Identity Verification" then ["identity_verification", "DIGITAL IDENTITY VERIFICATION"]
+                              when "Document Verification" then ["document_verification", "DOCUMENT VERIFICATION"]
+                              when "KYC Verification"      then ["kyc_verification", "KYC ACCOUNT OPENING"]
+                              when "SCRA Monitoring"       then ["scra_monitoring", "SCRA MONITORING"]
+                              when "Group Verification"    then ["group_verification", "GROUP AFFILIATION VERIFICATION"]
+                              when "Military"              then ["military", "MILITARY VERIFICATION"]
+                              when "Students"              then ["students", "STUDENT VERIFICATION"]
+                              when "Teachers"              then ["teachers", "TEACHER VERIFICATION"]
+                              when "First Responders"      then ["first_responders", "FIRST RESPONDER VERIFICATION"]
+                              when "Government Employees"  then ["government_employees", "GOVERNMENT EMPLOYEE VERIFICATION"]
+                              when "Government"            then ["government", "GOVERNMENT"]
+                              when "Healthcare"            then ["healthcare", "HEALTHCARE"]
+                              when "Financial Services"    then ["financial_services", "FINANCIAL SERVICES"]
+                              when "Ecommerce"             then ["ecommerce", "ECOMMERCE"]
+                              when "Ecommerce"             then ["ecommerce", "ECOMMERCE"]
+                              when "Request a Demo"        then ["request_demo", "CONTACT US"]
+                              else fail ("#{dropdown_link} not recognized")
+                              end
+
+  expect(page).to have_current_path(FigNewton.wallet.business_and_gov.send("#{expected_url}"),:url => true)
+
+  if dropdown_link == "Government"
+    expect(page).to have_css(".page-container-header > div > div > .title",:text => page_header)
+  else
+    expect(page).to have_css(".page-container-header > div > .title",:text => page_header)
+  end
+
+  step 'I visit "WalletHomepage"'
+end
+
+Given("I verify the For Individuals navigation dropdown link {string}") do |dropdown_link|
+  @WalletHomepage.click_shared_navlink_forindividuals
+  find("#js_dropdown2").click_link(dropdown_link)
+
+  expected_url, page_header = case dropdown_link
+                              when "Accessing Government Services"  then ["accessing_gov_services", "ACCESSING GOVERNMENT SERVICES"]
+                              when "Qualifying for Group Discounts" then ["group_discounts", "QUALIFYING FOR GROUP DISCOUNTS"]
+                              when "Shop Now"                       then ["shop_now", nil]
+                              else fail ("#{dropdown_link} not recognized")
+                              end
+
+  expect(page).to have_current_path(FigNewton.wallet.individuals.send("#{expected_url}"),:url => true)
+
+  unless dropdown_link == "Shop Now"
+    expect(page).to have_css(".page-container-header > div > .title",:text => page_header)
+  end
+
+  step 'I visit "WalletHomepage"'
+end
+
+Given(/^I verify the wallet My Account links direct to idp$/) do
+  @WalletHomepage.click_my_account_button
   expect(page).to have_current_path(FigNewton.idp.new_session_url,:url => true)
   return_to_previous_page
-end
-
-Given(/^I verify the wallet sign up links direct to idp$/) do
-  %w[click_shared_nav_sign_up_button click_hero_sign_up_button click_portable_sign_up_button click_home_cta_sign_up_button].each do |method|
-    @WalletHomepage.send(method)
-    expect(page).to have_current_path(FigNewton.idp.new_registration_url,:url => true)
-    return_to_previous_page
-  end
-end
-
-Given(/^I verify the Wallet shared navigation tab "([^"]*)" successfully links$/) do |nav_tab|
-  nav_link_url = case nav_tab
-                  when "About"    then FigNewton.idme.about_page
-                  when "Business" then FigNewton.idme.business_page
-                  when "Shop"     then FigNewton.wallet.shop
-                  when "Wallet"   then FigNewton.wallet.homepage.url
-                 end
-
-  if nav_tab == "Business"
-    business_window = window_opened_by do
-      find(".wallet-header-nav a",:text => nav_tab ).click
-    end
-    switch_to_window(business_window)
-    expect(page).to have_current_path(nav_link_url, :url => true)
-    close_current_browser
-    use_last_browser_created
-  end
 end